+++ src/Orm/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added support for an `options` key in the `OrmFactory::connection()` configuration array, forwarded as the driver-specific PDO connection options to the created `Connection`
 
+## [1.4.1] - 2026-07-31
+
+_No changes in this release._
+
 ## [1.4.0] - 2026-06-30
 
 ### Changed